+++ genericf2py/main.py
@@ -35,6 +35,10 @@ import sys
 import os
 here = os.path.abspath(os.path.dirname(__file__))
 
+# # python >=3.8 needs to be given permission to import DLL files.
+from genericf2py.find_mingw_lib import add_mingw_lib
+add_mingw_lib()
+
 
 # for multi-file projects see LICENSE file for authorship info
 # for single file projects, insert following information